Uniform scale for all Braille glyphs
[why]
All Braille glyphs shall keep their relative size and position, so that
we see a difference if a single dot is low or high for example.

[how]
Set the ScaleGlyph, so that the total bounding box of all glyphs is
scaled and shifted, but all individual glyphs with the same values (thus
keeping them consistent).

Also set 'Exact', as that is the 'spirit' used here. Exact means that
gaps in the symbol font are NOT skipped over, keeping codepoints as they
are. The Braille source font does not have gaps, but anyhow, we do not
want any 'codepoint compression' and express this now.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>