font-patcher: Preserve padding around Braille glyphs
[why]
Usually we scale symbols so that they are maximized within one 'cell' of
the font. If we do this for the Braille glyphs the bottom dots from un
upper line will (almost) touch the top dots of the next line. This is
not useful.

[how]
We add padding around the Braille glyphs, to get about the same distance
to the cell borders as the different dots in one glyph have to the
other (~10 %, 5% on each side).

This also activated the previous commit's shift-fix-code that is only
active on padded glyphs.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>